Settings: some fixes

+++ AUHTC/AUHTC/Model/Constants.cs
@@ -8,12 +8,12 @@ namespace AUHTC.Model
 {
     public class Constants
     {
-        public int MapHeight { get; set; }
+        public double MapHeight { get; set; }
 
-        public int MapWidth { get; set; }
+        public double MapWidth { get; set; }
 
-        public int MapTop { get; set; }
+        public double MapTop { get; set; }
 
-        public int MapLeft { get; set; }
+        public double MapLeft { get; set; }
     }
 }
